Use synchronous operations on ZipArchive in CoreCLR
