+++ src/Microsoft.Framework.PackageManager/Restore/NuGet/PackageUtilities.cs
@@ -2,7 +2,9 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
+using System.IO;
 using System.IO.Compression;
+using System.Threading.Tasks;
 
 namespace Microsoft.Framework.PackageManager
 {
@@ -19,6 +21,44 @@ namespace Microsoft.Framework.PackageManager
             }
 
             return null;
+        }
+
+        internal static async Task<Stream> OpenNuspecStreamFromNupkgAsync(PackageInfo package,
+            Func<PackageInfo, Task<Stream>> openNupkgStreamAsync,
+            IReport report)
+        {
+            using (var nupkgStream = await openNupkgStreamAsync(package))
+            {
+                try {
+                    using (var archive = new ZipArchive(nupkgStream, ZipArchiveMode.Read, leaveOpen: true))
+                    {
+                        var entry = archive.GetEntryOrdinalIgnoreCase(package.Id + ".nuspec");
+                        using (var entryStream = entry.Open())
+                        {
+                            var nuspecStream = new MemoryStream((int)entry.Length);
+#if ASPNETCORE50
+                            // System.IO.Compression.DeflateStream throws exception when multiple
+                            // async readers/writers are working on a single instance of it
+                            entryStream.CopyTo(nuspecStream);
+#else
+                            await entryStream.CopyToAsync(nuspecStream);
+#endif
+                            nuspecStream.Seek(0, SeekOrigin.Begin);
+                            return nuspecStream;
+                        }
+                    }
+                }
+                catch (InvalidDataException)
+                {
+                    var fileStream = nupkgStream as FileStream;
+                    if (fileStream != null)
+                    {
+                        report.WriteLine("The ZIP archive {0} is corrupt",
+                            fileStream.Name.Yellow().Bold());
+                    }
+                    throw;
+                }
+            }
         }
     }
 }
